requirements.txt
APScheduler==3.6.3 #
python-telegram-bot==13.4.1   #
pytz==2021.1 #
tornado==6.1 #
tzlocal==2.1 #